refactor(studio): remove unused types, ensure filter form inputs are focused first
