+++ packages/sanity/src/core/studio/components/navbar/search/SearchDialog.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components'
 import {useColorScheme} from '../../../colorScheme'
 import {RecentSearches} from './components/recentSearches/RecentSearches'
 import {SearchHeader} from './components/SearchHeader'
-import {SearchResults} from './components/SearchResults'
+import {SearchResults} from './components/searchResults/SearchResults'
 import {DocumentTypesVirtualList} from './components/filters/documentTypes/DocumentTypesVirtualList'
 import {CommandListProvider} from './contexts/commandList'
 import {useSearchState} from './contexts/search/useSearchState'
