+++ packages/sanity/src/core/studio/components/navbar/search/components/filters/documentTypes/items/TypeFilterItem.tsx
