--- packages/sanity/src/core/studio/components/navbar/search/components/searchResultItem/index.ts
@@ -1,1 +1,0 @@
-export * from './SearchResultItem'
