--- packages/sanity/src/core/studio/components/navbar/search/components/searchResultItem/types.ts
@@ -1,7 +1,0 @@
-import type {SanityDocument} from '@sanity/types'
-
-export interface SearchItemPreviewState {
-  draft?: SanityDocument | null
-  isLoading?: boolean
-  published?: SanityDocument | null
-}
