+++ packages/sanity/src/core/studio/components/navbar/search/components/searchResults/items/SearchResultItem.tsx
@@ -2,12 +2,12 @@ import {Box, ResponsivePaddingProps} from '@sanity/ui'
 import React, {forwardRef, MouseEvent, useMemo} from 'react'
 import type {VirtualItem} from '@tanstack/react-virtual'
 import styled from 'styled-components'
-import {PreviewCard} from '../../../../../../components/PreviewCard'
-import {FIXME} from '../../../../../../FIXME'
-import {useSchema} from '../../../../../../hooks'
-import type {WeightedHit} from '../../../../../../search'
-import {useDocumentPresence, useDocumentPreviewStore} from '../../../../../../store'
-import {getPublishedId} from '../../../../../../util/draftUtils'
+import {PreviewCard} from '../../../../../../../components/PreviewCard'
+import {FIXME} from '../../../../../../../FIXME'
+import {useSchema} from '../../../../../../../hooks'
+import type {WeightedHit} from '../../../../../../../search'
+import {useDocumentPresence, useDocumentPreviewStore} from '../../../../../../../store'
+import {getPublishedId} from '../../../../../../../util/draftUtils'
 import {DebugOverlay} from './DebugOverlay'
 import SearchResultItemPreview from './SearchResultItemPreview'
 import {IntentLink} from 'sanity/router'
